🩹 [FE] 특정 보드의 핀 목록을 전역 변수로 설정하는 코드 수정
