+++ frontend/assets/js/maps/board.js
@@ -1,6 +1,6 @@
 import { getBoardRequest, boardSimpleSaveRequest, getLoginUserInfoRequest, findBoardRequest, pinDeleteRequest } from "../request/content.js";
 import { $boardAddModal, $boardAddModalContent, $mainBoard, MARKERS, MAP, MY_BOARDS, PIN_SAVE_OVERLAY, PIN_SAVE_OVERLAY_CONTENT, ACCOUNT } from "./data.js";
-import { registerMainPin, pinSimpleSave } from "./pin.js";
+import { setMarkersFromServer, pinSimpleSave } from "./pin.js";
 
 // 타이틀, 태그로 보드 생성
 export async function boardSimpleSave(title, tags) {
@@ -29,119 +29,6 @@ export async function setMyBoard() {
 
     boards.forEach(board=>{
         MY_BOARDS.push(board);
-    });
-}
-
-// 핀 생성 추가 오버레이
-export function displayBoardsOnOverlay(markerInfo) {
-
-    if (MY_BOARDS === null || MY_BOARDS.length === 0) {
-        return;
-    }
-
-    MY_BOARDS.forEach(board=>{
-        let boardBox = document.createElement('div');
-        let titleBox = document.createElement('div');
-        let thumbnail = document.createElement('img');
-        let title = document.createElement('span');
-        let pinSaveBtn = document.createElement('div');
-
-        pinSaveBtn.innerText = '생성';
-        pinSaveBtn.classList.add('pin_save_btn');
-        let pinsaved = false;
-
-        // 해당 핀이 보드에 생성된 경우 '생성됨' 처리
-        for (let i=0; i<board.pins.length; i++) {
-            if (board.pins[i] == markerInfo.id) {
-                pinSaveBtn.innerText = '생성됨';
-                pinSaveBtn.classList.remove('pin_save_btn');
-                pinSaveBtn.classList.add('pin_saved_btn');
-                pinsaved = true;
-                break;
-            }
-        }
-
-        if (board.thumbnail_imgs === null || board.thumbnail_imgs === undefined || board.thumbnail_imgs.length === 0) {
-            thumbnail.src = 'assets/img/favspot.png';
-        } else {
-            for (let i=0; i<board.thumbnail_imgs.length; i++) {
-                if(board.thumbnail_imgs[i] === null || board.thumbnail_imgs[i] === undefined || board.thumbnail_imgs[i].length > 0) {
-                    thumbnail.src =  board.thumbnail_imgs[i];
-                    break;
-                }
-            }
-        }
-
-        title.innerText = board.title;
-        thumbnail.style.width = '50px';
-        thumbnail.style.height = '40px';
-        thumbnail.style.marginRight = '7px';
-        thumbnail.style.borderRadius = '10px';
-
-        titleBox.appendChild(thumbnail);
-        titleBox.append(title);
-        boardBox.appendChild(titleBox);
-        boardBox.appendChild(pinSaveBtn);
-      
-        boardBox.classList.add('pin_save_board');
-
-        // 생성 버튼 클릭 이벤트
-        pinSimpleSaveEvent(pinSaveBtn, board, markerInfo, pinsaved);
-
-        PIN_SAVE_OVERLAY_CONTENT.appendChild(boardBox);
-    });
-
-    // 보드 만들기 버튼/이벤트 추가
-    let boardAddBtnBox = document.createElement('div');
-    let boardAddBtn = document.createElement('div');
-    
-    boardAddBtn.innerText = '보드 만들기';
-    boardAddBtnBox.classList.add('board_add_box');
-
-    boardAddBtnBox.addEventListener('click', ()=>{
-
-        if (!ACCOUNT.login) {
-            alert('로그인이 필요합니다')
-            return;
-        }
-
-        $boardAddModal.style.display = 'flex';
-        $boardAddModalContent.style.display = 'flex';
-    });
-
-    boardAddBtnBox.appendChild(boardAddBtn);
-    PIN_SAVE_OVERLAY_CONTENT.appendChild(boardAddBtnBox);
-                  
-    PIN_SAVE_OVERLAY.setContent(PIN_SAVE_OVERLAY_CONTENT);
-    PIN_SAVE_OVERLAY.setPosition(markerInfo.position);
-    PIN_SAVE_OVERLAY.setMap(MAP);
-    PIN_SAVE_OVERLAY.setVisible(true);
-}
-
-// 생성 버튼 클릭 이벤트
-function pinSimpleSaveEvent(element, board, place, pinsaved) {
-
-    element.addEventListener('click', async ()=>{
-        if (!pinsaved) {
-            let saveSucceess = pinSimpleSave(board, place);
-            if (saveSucceess) {
-                setMyBoard();
-                element.innerText = '생성됨';
-                element.classList.remove('pin_save_btn');
-                element.classList.add('pin_saved_btn');
-                pinsaved = true;
-                return;
-            }
-        } else {
-            if (confirm('핀을 삭제하시겠습니까?')) {
-                // let response = await pinDeleteRequest(place.id);
-
-                element.innerText = '생성';
-                element.classList.remove('pin_saved_btn');
-                element.classList.add('pin_save_btn');
-                pinsaved = false;
-            }
-        }
     });
 }
 
@@ -215,8 +102,7 @@ export function displayMainBoards(boards) {
             if (response.status >= 400 && response.status < 600) {
                 return;
             }
-            let pins = JSON.stringify(response.pins);
-            window.localStorage.setItem('pins', pins);
+            window.localStorage.setItem('BOARD_ID', randomBoards[i].id);
             // location.href = '';
         });
